requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
pymysql==1.1.0
cryptography==41.0.7
matplotlib==3.8.2
python-dotenv==1.0.1